+++ emulator/commands/c_mov_bxv.hpp
@@ -19,7 +19,7 @@ public:
 
     int run(std::vector<uint8_t> args, Runtime* runtime) override // mov_bxv <v²>
     {
-        uint8_t v = ppt8_std::combine_8bit_to_16bit(args[0], args[1]);
+        uint16_t v = ppt8_std::combine_8bit_to_16bit(args[0], args[1]);
         runtime->getBX()->setValueComplete16bit(v);
         return 0;
     }
