+++ emulator/commands/c_mov_cxv.hpp
@@ -19,7 +19,7 @@ public:
 
     int run(std::vector<uint8_t> args, Runtime* runtime) override // mov_cxv <v²>
     {
-        uint8_t v = ppt8_std::combine_8bit_to_16bit(args[0], args[1]);
+        uint16_t v = ppt8_std::combine_8bit_to_16bit(args[0], args[1]);
         runtime->getCX()->setValueComplete16bit(v);
         return 0;
     }
